requirements.txt
load_dotenv
requests
cohere
bs4
termcolor
streamlit